rename antismash files with unique names
