+++ tests/genomics/antismash/test_antismash_loader.py
